backend/requirements.txt
anthropic>=0.50.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
Flask==3.1.3
flask-marshmallow==1.4.0
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.3.1
Flask-Cors==5.0.1
greenlet==3.3.2
gunicorn==23.0.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
packaging==26.1
Pillow>=11.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
Pygments==2.20.0
pytest==9.0.3
pytest-flask==1.3.0
python-dotenv==1.0.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
supabase>=2.0
typing_extensions==4.15.0
Werkzeug==3.1.6
